Changes:  - fixes for building against external Boost  - fixes for compiler warnings (gcc signed/unsigned mismatch)
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@4989 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ libraries/utilities/include/mcrl2/utilities/font_renderer.h
@@ -32,17 +32,17 @@ namespace mcrl2
     /**
      * \brief Number of characters within the character set.
      **/
-    const int CHARSETSIZE = 96;
+    const size_t CHARSETSIZE = 96;
 
     /**
      * \brief Pixel width per character.
      **/
-    const int CHARWIDTH = 16;
+    const size_t CHARWIDTH = 16;
 
     /**
      * \brief Pixel height per character.
      **/
-    const int CHARHEIGHT = 32;
+    const size_t CHARHEIGHT = 32;
 
     /**
      * \brief A font renderer for use with OpenGL
